# src/Ai.Tlbx.MidTerm/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "MidTerm API",
    "description": "MidTerm terminal multiplexer API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sessions": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/CreateSessionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/attention": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "agentOnly",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionAttentionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workers/bootstrap": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkerBootstrapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerBootstrapResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/reorder": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}": {
      "delete": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/resize": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResizeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/state": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeBuffer",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "includeBufferBase64",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/input": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "byte"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/input/text": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/input/keys": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionKeyInputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/input/prompt": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionPromptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/buffer": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/buffer/text": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeBase64",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionBufferTextResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/buffer/tail": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lines",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 120
            }
          },
          {
            "name": "stripAnsi",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/activity": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "seconds",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 120
            }
          },
          {
            "name": "bellLimit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionActivityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/agent": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tailLines",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 80
            }
          },
          {
            "name": "activitySeconds",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 90
            }
          },
          {
            "name": "bellLimit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSessionVibeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/agent/feed": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tailLines",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 80
            }
          },
          {
            "name": "activitySeconds",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 90
            }
          },
          {
            "name": "bellLimit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 8
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSessionFeedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/name": {
      "put": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "auto",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/notes": {
      "put": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSessionNotesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/topic": {
      "put": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSessionTopicRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/control": {
      "put": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSessionControlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/upload": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/paste-clipboard-image": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/inject-guidance": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InjectGuidanceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/bookmark": {
      "put": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBookmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "required": [
          "cols",
          "rows"
        ],
        "type": "object",
        "properties": {
          "cols": {
            "type": "integer",
            "format": "int32"
          },
          "rows": {
            "type": "integer",
            "format": "int32"
          },
          "shell": {
            "type": [
              "null",
              "string"
            ]
          },
          "workingDirectory": {
            "type": [
              "null",
              "string"
            ]
          },
          "spaceId": {
            "type": [
              "null",
              "string"
            ]
          },
          "workspacePath": {
            "type": [
              "null",
              "string"
            ]
          },
          "surface": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "SessionAttentionResponse": {
        "required": [
          "generatedAt",
          "agentOnly",
          "attentionCount",
          "sessions"
        ],
        "type": "object",
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "agentOnly": {
            "type": "boolean"
          },
          "attentionCount": {
            "type": "integer",
            "format": "int32"
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionAttentionItem"
            }
          }
        }
      },
...
      "SessionInfoDto": {
        "required": [
          "id",
          "pid",
          "createdAt",
          "isRunning",
          "exitCode",
          "cols",
          "rows",
          "shellType",
          "name",
          "terminalTitle",
          "topic",
          "notes",
          "manuallyNamed",
          "currentDirectory",
          "foregroundPid",
          "foregroundName",
          "foregroundCommandLine",
          "foregroundDisplayName",
          "foregroundProcessIdentity",
          "agentAttachPoint",
          "order",
          "parentSessionId",
          "bookmarkId",
          "spaceId",
          "workspacePath",
          "surface",
          "isAdHoc",
          "agentControlled",
          "appServerControlOnly",
          "profileHint",
          "appServerControlResumeThreadId",
          "hasAppServerControlHistory",
          "supervisor"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pid": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isRunning": {
            "type": "boolean"
          },
          "exitCode": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "cols": {
            "type": "integer",
            "format": "int32"
          },
          "rows": {
            "type": "integer",
            "format": "int32"
          },
          "shellType": {
            "type": "string"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "terminalTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "topic": {
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "type": [
              "null",
              "string"
            ]
          },
          "manuallyNamed": {
            "type": "boolean"
          },
          "currentDirectory": {
            "type": [
              "null",
              "string"
            ]
          },
          "foregroundPid": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "foregroundName": {
            "type": [
              "null",
              "string"
            ]
          },
          "foregroundCommandLine": {
            "type": [
              "null",
              "string"
            ]
          },
          "foregroundDisplayName": {
            "type": [
              "null",
              "string"
            ]
          },
          "foregroundProcessIdentity": {
            "type": [
              "null",
              "string"
            ]
          },
          "agentAttachPoint": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SessionAgentAttachPoint"
              }
            ]
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "parentSessionId": {
            "type": [
              "null",
              "string"
            ]
          },
          "bookmarkId": {
            "type": [
              "null",
              "string"
            ]
          },
          "spaceId": {
            "type": [
              "null",
              "string"
            ]
          },
          "workspacePath": {
            "type": [
              "null",
              "string"
            ]
          },
          "surface": {
            "type": [
              "null",
              "string"
            ]
          },
          "isAdHoc": {
            "type": "boolean"
          },
          "agentControlled": {
            "type": "boolean"
          },
          "appServerControlOnly": {
            "type": "boolean"
          },
          "profileHint": {
            "type": [
              "null",
              "string"
            ]
          },
          "appServerControlResumeThreadId": {
            "type": [
              "null",
              "string"
            ]
          },
          "hasAppServerControlHistory": {
            "type": "boolean"
          },
          "supervisor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SessionSupervisorInfoDto"
              }
            ]
          }
        }
      },
...
      "SessionListDto": {
        "required": [
          "sessions"
        ],
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionInfoDto"
            }
          }
        }
      },
...
    }
  },
...
}